add wage data to repo for consistency, adding and adjusting cols based on 20230517 convo
